services/journal-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
prometheus-client